requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.10
    # via openai
aiosignal==1.3.1
    # via aiohttp
attrs==24.2.0
    # via aiohttp
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via -r requirements.in
decorator==4.4.2
    # via moviepy
ffmpeg-python==0.2.0
    # via -r requirements.in
filelock==3.16.1
    # via torch
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.10.0
    # via torch
future==1.0.0
    # via ffmpeg-python
idna==3.10
    # via
    #   requests
    #   yarl
imageio==2.36.1
    # via moviepy
imageio-ffmpeg==0.5.1
    # via moviepy
jinja2==3.1.4
    # via torch
llvmlite==0.43.0
    # via numba
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.5.0
    # via openai-whisper
moviepy==1.0.3
    # via -r requirements.in
mpmath==1.3.0
    # via sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
networkx==3.4.2
    # via torch
numba==0.60.0
    # via openai-whisper
numpy==2.0.2
    # via
    #   imageio
    #   moviepy
    #   numba
    #   openai-whisper
openai==0.28.0
    # via -r requirements.in
openai-whisper==20240930
    # via -r requirements.in
pillow==10.4.0
    # via imageio
proglog==0.1.10
    # via moviepy
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
pydub==0.25.1
    # via -r requirements.in
pygments==2.18.0
    # via rich
python-dotenv==1.0.1
    # via -r requirements.in
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   moviepy
    #   openai
    #   tiktoken
rich==13.9.4
    # via -r requirements.in
sympy==1.13.1
    # via torch
tiktoken==0.8.0
    # via openai-whisper
torch==2.5.1
    # via openai-whisper
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   moviepy
    #   openai
    #   openai-whisper
    #   proglog
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   torch
urllib3==2.2.3
    # via requests
yarl==1.18.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
